Move to using CMake
