+++ .gitignore
@@ -34,6 +34,6 @@
 .DS_Store
 *.sublime-workspace
 
-Bach/bin
+build
 Bach/geom/Geom_output.xml
 *.root
