+++ TbKernel/include/TbClipboard.h
